Resolve the ABI tag dynamically as well
